move signature to traits
