+++ lang/Cargo.toml
@@ -20,7 +20,6 @@ ink = { version = "4.2.1", default-features = false}
 ink_ir = { version = "4.2.1", default-features = false }
 scale = { package = "parity-scale-codec", version = "3", default-features = false, features = ["derive"] }
 scale-info = { version = "2.6", default-features = false, features = ["derive"] }
-blake2 = { version = "0.10.6", default-features = false }
 
 xxhash-rust = { version = "0.8", features = ["const_xxh32"] }
 const_format = "0.2.21"
