+++ lang/codegen/Cargo.toml
@@ -22,10 +22,10 @@ fs2 = "0.4.3"
 serde = { version = "1.0", features = ["derive"] }
 cargo_metadata = "0.13.1"
 unwrap = "1.2.1"
-blake2 = "0.9"
+blake2 = "0.10.6"
 heck = "0.3.1"
-ink_ir = { version = "4.1.0-beta", default-features = false }
-ink_primitives = { version = "4.1.0-beta", default-features = false }
+ink_ir = { version = "4.2.1", default-features = false }
+ink_primitives = { version = "4.2.1", default-features = false }
 synstructure = "0.12"
 
 [lib]
